requirements.txt
flask
numpy
opencv-python
tensorflow
Pillow==8.3.2